requirements.txt
black==22.10.0
click==8.1.3
delta-spark==2.1.1
importlib-metadata==5.0.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.3
py4j==0.10.9.5
pyspark==3.3.1
PyYAML==6.0
regex==2022.10.31
tomli==2.0.1
typer==0.7.0
typing_extensions==4.4.0
yetl-framework==0.0.25.dev2
zipp==3.10.0
